Refact: move resources to resources pckg
