+++ server/core/resources/auth/auth.router.js
