+++ server/core/resources/visualization/vis.model.js
