+++ server/core/resources/visualization/vis.router.js
