+++ server/core/router.js
@@ -1,14 +1,10 @@
-/**
- * Api routes.
- */
-
 const path = require("path");
 const express = require("express");
 const endpoint = require("../constants/endpoint");
 
-const visRoutes = require("./visualization/vis.router");
-const userRoutes = require("./user/user.router");
-const authRoutes = require("./auth/auth.router");
+const visRoutes = require("./resources/visualization/vis.router");
+const userRoutes = require("./resources/user/user.router");
+const authRoutes = require("./resources/auth/auth.router");
 
 const STATIC_FILES_DIRECTORY = path.join(__dirname, "../../static");
